+++ pacman_mirrors/configfn.py
@@ -1,0 +1,75 @@
+#!/usr/bin/env python
+#
+# This file is part of pacman-mirrors.
+#
+# pacman-mirrors is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# pacman-mirrors is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with pacman-mirrors.  If not, see <http://www.gnu.org/licenses/>.
+#
+# Authors: Frede Hundewadt <frede@hundewadt.dk>
+
+"""Pacman-Mirrors Configuration Functions"""
+
+from . import txt
+from .configuration import CONFIG_FILE, MIRROR_DIR, MIRROR_FILE, MIRROR_LIST
+
+
+class ConfigFn:
+    """Configuration Functions"""
+
+    @staticmethod
+    def build_config():
+        """Get config informations"""
+        # initialising defaults
+        # information which can differ from these defaults
+        # is fetched from config file
+        config = {
+            "branch": "stable",
+            "config_file": CONFIG_FILE,
+            "method": "rank",
+            "mirror_dir": MIRROR_DIR,
+            "mirror_file": MIRROR_FILE,
+            "mirror_list": MIRROR_LIST,
+            "no_update": False,
+            "only_country": [],
+        }
+        try:
+            # read configuration from file
+            with open(config["config_file"]) as conf:
+                for line in conf:
+                    line = line.strip()
+                    if line.startswith("#") or "=" not in line:
+                        continue
+                    (key, value) = line.split("=", 1)
+                    key = key.rstrip()
+                    value = value.lstrip()
+                    if key and value:
+                        if value.startswith("\"") and value.endswith("\""):
+                            value = value[1:-1]
+                        if key == "Method":
+                            config["method"] = value
+                        elif key == "Branch":
+                            config["branch"] = value
+                        elif key == "OnlyCountry":
+                            config["only_country"] = value.split(",")
+                        elif key == "MirrorlistsDir":
+                            config["mirror_dir"] = value
+                        elif key == "OutputMirrorlist":
+                            config["mirror_list"] = value
+                        elif key == "NoUpdate":
+                            config["no_update"] = value
+        except (PermissionError, OSError) as err:
+            print(".: {} {}: {}: {}".format(txt.ERR_CLR,
+                                            txt.CANNOT_READ_FILE,
+                                            err.filename,
+                                            err.strerror))
+        return config